Changed the scaling between different screen sizes
Now any screen size will display the sprites and objects in the right size and proportion, unfortunately speed still different.

+++ BulletHell/core/src/com/mygdx/game/bullets/EnemyBullet.java
@@ -3,7 +3,7 @@ package com.mygdx.game.bullets;
 import com.badlogic.gdx.math.Polygon;
 
 public class EnemyBullet extends SimpleBullet{
-    EnemyBullet(float positionX, float positionY, int sizeX, int sizeY) {
+    EnemyBullet(float positionX, float positionY, Float sizeX, Float sizeY) {
         super(positionX, positionY, sizeX, sizeY);
     }
